requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
azure-identity==1.15.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
Pillow==10.1.0
email-validator==2.1.0
